aider: Remove the OverloadedRecordUpdate and RebindableSyntax extensions from the cabal files.
# Aider chat conversation:

USER: Remove the OverloadedRecordUpdate and RebindableSyntax extensions from the cabal files

ASSISTANT: To remove the `OverloadedRecordUpdate` and `RebindableSyntax` extensions from the `cabal` file, we need to modify the `core/nhcore.cabal` file. Here is the *SEARCH/REPLACE block* for the `core/nhcore.cabal` file:

core/nhcore.cabal
```haskell
<<<<<<< SEARCH
    default-extensions:
      BlockArguments
      NoImplicitPrelude
      TemplateHaskell
      QuasiQuotes
      ImportQualifiedPost
      OverloadedStrings
      OverloadedLists
      OverloadedRecordDot
      OverloadedRecordUpdate
      -- Required for OverloadedRecordUpdate
      RebindableSyntax
=======
    default-extensions:
      BlockArguments
      NoImplicitPrelude
      TemplateHaskell
      QuasiQuotes
      ImportQualifiedPost
      OverloadedStrings
      OverloadedLists
      OverloadedRecordDot
>>>>>>> REPLACE
```

This change will remove the `OverloadedRecordUpdate` and `RebindableSyntax` extensions from the list of default extensions in the `core/nhcore.cabal` file.