aider: Remove `OverloadedRecordUpdate` and `RebindableSyntax` extensions from `cli/nhcli.cabal`.
# Aider chat conversation:

ASSISTANT: To remove the `OverloadedRecordUpdate` and `RebindableSyntax` extensions from the `cli/nhcli.cabal` file, we need to modify the file accordingly. Here is the *SEARCH/REPLACE block* for the `cli/nhcli.cabal` file:

cli/nhcli.cabal
```haskell
<<<<<<< SEARCH
    default-extensions:
      BlockArguments
      NoImplicitPrelude
      TemplateHaskell
      QuasiQuotes
      ImportQualifiedPost
      OverloadedStrings
      OverloadedLists
      OverloadedRecordDot
      OverloadedRecordUpdate
      -- Required for OverloadedRecordUpdate
      RebindableSyntax
=======
    default-extensions:
      BlockArguments
      NoImplicitPrelude
      TemplateHaskell
      QuasiQuotes
      ImportQualifiedPost
      OverloadedStrings
      OverloadedLists
      OverloadedRecordDot
>>>>>>> REPLACE
```

This change will remove the `OverloadedRecordUpdate` and `RebindableSyntax` extensions from the list of default extensions in the `cli/nhcli.cabal` file.